refactor(act): tighten internal/ barrel to actual external consumers
The internal/ barrel was re-exporting every symbol via `export *`, even
ones used only inside internal/ (drain ops, event-sourcing primitives,
schema/init merge helpers). Two issues:

1. Silent name-collision risk if two internal modules ever shared a name
2. Misleading surface — readers can't see what is actually consumed
   from outside without grepping

Tightened to four lines: DrainOps and EsOps types (used by act.ts),
_this_ + mergeProjection + registerState (used by builders), and
buildDrain + buildEs (used by act.ts). Drop the export keyword on
baseTypeName / mergeSchemas / mergeInits — they are merge.ts internals.

Test files (cache.spec.ts, cache.bench.ts, event-sourcing.spec.ts,
act-pg/cache.bench.ts) that imported `action`/`load`/`snap` from the
barrel now import directly from internal/event-sourcing.js — they
exercise the bare implementations and never used the barrel for
abstraction.
